refactor: split service config schemas
