+++ services/enqueue-api/config.py
@@ -1,0 +1,8 @@
+import os
+
+from constants import PUBSUB_TOPIC_DEFAULT, REGION_DEFAULT
+
+PROJECT_ID = os.getenv("PROJECT_ID")
+PUBSUB_TOPIC = os.getenv("PUBSUB_TOPIC", PUBSUB_TOPIC_DEFAULT)
+CLOUD_RUN_JOB_NAME = os.getenv("CLOUD_RUN_JOB_NAME")
+REGION = os.getenv("REGION", REGION_DEFAULT)
