+++ services/enqueue-api/constants.py
@@ -1,0 +1,4 @@
+PUBSUB_TOPIC_DEFAULT = "race-processing"
+REGION_DEFAULT = "us-central1"
+SERVICE_NAME = "smartervote-enqueue-api"
+SOURCE_NAME = "enqueue-api"
