+++ services/races-api/constants.py
@@ -1,0 +1,1 @@
+DEFAULT_DATA_DIR = "data/published/"
